feat: implement seller delete account page
